# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface CredentialsInput {
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface PasswordChangeInput {
  /** @minLength 1 */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  newPassword: string;
}

export interface PasswordResetRequestInput {
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
}

export interface PasswordResetCompleteInput {
  /** @minLength 20 */
  token: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  newPassword: string;
}

export interface LicenseInput {
  /**
     * @minimum 1
     * @nullable
     */
  days: number | null;
}

export interface TemplateInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 500 */
  description?: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence?: number;
  /**
     * @maxLength 14000000
     * @nullable
     */
  imageData?: string | null;
}

export interface TemplateUpdate {
  isActive: boolean;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  active: 'active',
  rejected: 'rejected',
  expired: 'expired',
} as const;

export interface User {
  id: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  expirationAt: string | null;
  createdAt: string;
}

export interface Template {
  id: string;
  name: string;
  description: string;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  imageData?: string | null;
  confidence: number;
  isActive: boolean;
}

export type AuditLogMetadata = { [key: string]: unknown };

export interface AuditLog {
  id: string;
  adminId: string;
  action: string;
  /** @nullable */
  targetUserId?: string | null;
  metadata: AuditLogMetadata;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface UserResponse {
  user: User;
}

export interface TemplateResponse {
  template: Template;
}

export interface MessageResponse {
  message: string;
}

export type ListUsers200 = {
  users: User[];
};

export type ListTemplates200 = {
  templates: Template[];
};

export type ListAuditLogs200 = {
  logs: AuditLog[];
};