# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Ultra AutoDetector account, license, and detection-template API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: users
  - name: templates
  - name: audit
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a pending account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      responses:
        "204":
          description: Signed out
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeInput"
      responses:
        "204":
          description: Password changed
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request password reset instructions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestInput"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/password-reset/complete:
    post:
      operationId: completePasswordReset
      tags: [auth]
      summary: Complete a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetCompleteInput"
      responses:
        "204":
          description: Password reset completed
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all accounts for administrators
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
  /users/{id}/license:
    patch:
      operationId: grantLicense
      tags: [users]
      summary: Grant or extend a user license
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/{id}/reject:
    post:
      operationId: rejectUser
      tags: [users]
      summary: Reject a pending account
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List active detection templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: object
                required: [templates]
                properties:
                  templates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a detection template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateResponse"
  /templates/{id}:
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Activate or deactivate a template
      parameters:
        - $ref: "#/components/parameters/TemplateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateResponse"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - $ref: "#/components/parameters/TemplateId"
      responses:
        "204":
          description: Deleted
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List recent administrator activity
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                required: [logs]
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLog"
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
    TemplateId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    CredentialsInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
        password:
          type: string
          minLength: 8
          maxLength: 128
    PasswordChangeInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8, maxLength: 128 }
    PasswordResetRequestInput:
      type: object
      required: [email]
      properties:
        email: { type: string, pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$" }
    PasswordResetCompleteInput:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string, minLength: 20 }
        newPassword: { type: string, minLength: 8, maxLength: 128 }
    LicenseInput:
      type: object
      required: [days]
      properties:
        days:
          type: [number, "null"]
          minimum: 1
    TemplateInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        imageData: { type: [string, "null"], maxLength: 14000000 }
    TemplateUpdate:
      type: object
      required: [isActive]
      properties:
        isActive: { type: boolean }
    User:
      type: object
      required: [id, email, role, status, expirationAt, createdAt]
      properties:
        id: { type: string }
        email: { type: string, pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$" }
        role: { type: string, enum: [user, admin] }
        status: { type: string, enum: [pending, active, rejected, expired] }
        expirationAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    Template:
      type: object
      required: [id, name, description, confidence, isActive]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        downloadUrl: { type: [string, "null"] }
        imageData: { type: [string, "null"] }
        confidence: { type: number }
        isActive: { type: boolean }
    AuditLog:
      type: object
      required: [id, adminId, action, metadata, createdAt]
      properties:
        id: { type: string }
        adminId: { type: string }
        action: { type: string }
        targetUserId: { type: [string, "null"] }
        metadata: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    TemplateResponse:
      type: object
      required: [template]
      properties:
        template:
          $ref: "#/components/schemas/Template"
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }