# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */
import type { AuditLogMetadata } from './auditLogMetadata';

export interface AuditLog {
  id: string;
  adminId: string;
  action: string;
  /** @nullable */
  targetUserId?: string | null;
  metadata: AuditLogMetadata;
  createdAt: Date;
}