# lib/api-zod/src/generated/types/licenseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */

export interface LicenseInput {
  /**
     * @minimum 1
     * @nullable
     */
  days: number | null;
}