# lib/api-zod/src/generated/types/passwordChangeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */

export interface PasswordChangeInput {
  /** @minLength 1 */
  currentPassword: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  newPassword: string;
}