# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */

export interface TemplateInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 500 */
  description?: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence?: number;
  /**
     * @maxLength 14000000
     * @nullable
     */
  imageData?: string | null;
}