# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ultra AutoDetector account, license, and detection-template API
 * OpenAPI spec version: 0.2.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: string;
  /** @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$ */
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  expirationAt: Date | null;
  createdAt: Date;
}